requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
email-validator
pydantic-settings